refactor: consolidate JSON utilities in @lobu/core
Move bigint-safe `toJsonSafe` and `parseJsonObject` from
packages/server/src/utils/json.ts into packages/core/src/utils/json.ts
and delete the server copy. Server importers now pull all four helpers
(safeJsonParse, safeJsonStringify, toJsonSafe, parseJsonObject) from
`@lobu/core`.
